--- src/main/java/com/example/todo/domain/exception/TodoNotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.todo.domain.exception;
-
-public class TodoNotFoundException extends RuntimeException {
-    
-    public TodoNotFoundException(Long id) {
-        super("Todo not found with id: " + id);
-    }
-}
