--- src/main/java/com/example/todo/domain/exception/TodoValidationException.java
@@ -1,8 +1,0 @@
-package com.example.todo.domain.exception;
-
-public class TodoValidationException extends RuntimeException {
-    
-    public TodoValidationException(String message) {
-        super(message);
-    }
-}
