+++ src/main/java/com/example/todo/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.example.todo.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponse {
+    private int status;
+    private String message;
+    private LocalDateTime timestamp;
+}
