--- src/main/java/com/example/todo/interfaces/exception/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.example.todo.interfaces.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponse {
-    private int status;
-    private String message;
-    private LocalDateTime timestamp;
-}
