+++ src/main/java/com/example/todo/todo/domain/exception/TodoNotFoundException.java
@@ -1,0 +1,8 @@
+package com.example.todo.todo.domain.exception;
+
+public class TodoNotFoundException extends RuntimeException {
+    
+    public TodoNotFoundException(Long id) {
+        super("Todo not found with id: " + id);
+    }
+}
