+++ src/main/java/com/example/todo/todo/domain/exception/TodoValidationException.java
@@ -1,0 +1,8 @@
+package com.example.todo.todo.domain.exception;
+
+public class TodoValidationException extends RuntimeException {
+    
+    public TodoValidationException(String message) {
+        super(message);
+    }
+}
