+++ src/main/java/com/example/todo/todo/infrastructure/persistence/JpaTodoRepository.java
@@ -1,0 +1,12 @@
+package com.example.todo.todo.infrastructure.persistence;
+
+import com.example.todo.todo.domain.model.Todo;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface JpaTodoRepository extends JpaRepository<Todo, Long> {
+    List<Todo> findByCompleted(boolean completed);
+}
