+++ src/main/java/com/example/todo/user/domain/model/Gender.java
@@ -1,0 +1,19 @@
+package com.example.todo.user.domain.model;
+
+/**
+ * 사용자 성별을 나타내는 enum
+ */
+public enum Gender {
+    MALE("남"),
+    FEMALE("여");
+
+    private final String displayName;
+
+    Gender(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+}
